+++ apps/cli/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @perceo/perceo
 
+## 0.3.5
+
+### Patch Changes
+
+- Try to fix the .env not working again.
+- Updated dependencies
+  - @perceo/observer-engine@2.0.5
+  - @perceo/supabase@0.2.4
+
+## 0.3.4
+
+### Patch Changes
+
+- Fix anon key and supabase url not present for login before pulling public_env
+- Updated dependencies
+  - @perceo/observer-engine@2.0.4
+  - @perceo/supabase@0.2.3
+
+## 0.3.3
+
+### Patch Changes
+
+- Adding loading for .env variables from public env
+- Updated dependencies
+  - @perceo/observer-engine@2.0.3
+  - @perceo/supabase@0.2.2
+
+## 0.3.2
+
+### Patch Changes
+
+- Fix a minor rpc error
+- Updated dependencies
+  - @perceo/supabase@0.2.1
+  - @perceo/observer-engine@2.0.2
+
 ## 0.3.1
 
 ### Patch Changes